Add new ShipExtensions file
This file hosts a few utility functions that were lacking in both MathUtil
and VectorUtil. All of these methods added to the ShipAPIExtensions file are
extension functions to the ShipAPI class.

The added functions are:
- getForwardVector() - returns a Vector2f of where the (calling) ship is facing
- getAngleToAnotherShip(ShipAPI anotherShip) - returns the angle between the (calling) ship and *anotherShip*
- getAbsoluteAngleToAnotherShip(ShipAPI anotherShip) - similar to previous one, except it returns an absolute angle

While I do understand that getForwardVector() is very similar to calling Misc.getUnitVectorAtDegreeAngle(ship.getFacing())
it's just so ergonomically better than having to repeat that every time you need a ship's forward vector.