Add KotlinExtensions while at it
This file hosts generic kotlin extensions. In this particular case, the only method
added to it is the safeEquals() method, which is a null-safe way of testing equality
(by delegating to the objects' equals() method) when either or both of the objects are null.

As more useful things show up, more will be added.